pdsi_FourCastNet/requirements.txt
# --------- pytorch --------- #
torch==1.10.1
pytorch-lightning==1.9.1
torchmetrics>=0.7.0
torcheval==0.0.7

# --------- modulus --------- #
nvidia-modulus-sym==1.3.0

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis
nbstripout      # remove output from jupyter notebooks

# --------- other --------- #
python-dotenv   # loading env variables from .env file
rich            # beautiful text formatting in3 terminal
pytest          # tests
sh              # for running bash commands in some tests
pudb            # debugger
seaborn>=0.10.1 # plotting utils
tqdm
